Skriver om en funksjon til en komponent slik at man følger den nye standarden til react-router
